Move LogAnalytics.groovy to model.config
