+++ src/test/groovy/org/prebid/server/functional/model/config/AnalyticsModule.groovy
@@ -1,7 +1,6 @@
 package org.prebid.server.functional.model.config
 
 import groovy.transform.ToString
-import org.prebid.server.functional.tests.LogAnalytics
 
 @ToString(includeNames = true, ignoreNulls = true)
 class AnalyticsModule {
