+++ src/test/groovy/org/prebid/server/functional/model/request/auction/PrebidAnalytics.groovy
@@ -1,7 +1,7 @@
 package org.prebid.server.functional.model.request.auction
 
 import groovy.transform.ToString
-import org.prebid.server.functional.tests.LogAnalytics
+import org.prebid.server.functional.model.config.LogAnalytics
 
 @ToString(includeNames = true, ignoreNulls = true)
 class PrebidAnalytics {
